4. Production/deploy/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.7.1            # via tb-nightly, tensorflow
appdirs==1.4.3            # via black
astor==0.8.0              # via tensorflow
astroid==2.2.5            # via pylint
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, flake8-bugbear, packaging, pytest
black==19.3b0
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
click==7.0                # via black, flask
coverage==4.5.4           # via pytest-cov
entrypoints==0.3          # via flake8
flake8-bugbear==19.8.0
flake8==3.7.8
flask==1.1.1
gast==0.2.2               # via tensorflow
google-api-core==1.14.2   # via google-cloud-core
google-auth==1.6.3        # via google-api-core, google-cloud-storage
google-cloud-core==1.0.3  # via google-cloud-storage
google-cloud-storage==1.18.0
google-pasta==0.1.7       # via tensorflow
google-resumable-media==0.3.2  # via google-cloud-storage
googleapis-common-protos==1.6.0  # via google-api-core
grpcio==1.23.0            # via tb-nightly, tensorflow
h5py==2.9.0               # via keras-applications
idna==2.8                 # via requests
importlib-metadata==0.19  # via pluggy, pytest
isort==4.3.21
itsdangerous==1.1.0       # via flask
jinja2==2.10.1            # via flask
keras-applications==1.0.8  # via tensorflow
keras-preprocessing==1.1.0  # via tensorflow
lazy-object-proxy==1.4.1  # via astroid
markdown==3.1.1           # via tb-nightly
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8, pylint
more-itertools==7.2.0     # via pytest
mypy-extensions==0.4.1    # via mypy
mypy==0.720
numpy==1.17.0
packaging==19.1           # via pytest
pillow==6.1.0
pluggy==0.12.0            # via pytest
protobuf==3.9.1           # via google-api-core, googleapis-common-protos, tb-nightly, tensorflow
py==1.8.0                 # via pytest
pyasn1-modules==0.2.6     # via google-auth
pyasn1==0.4.6             # via pyasn1-modules, rsa
pycodestyle==2.5.0        # via flake8
pydocstyle==4.0.1
pyflakes==2.1.1           # via flake8
pylint==2.3.1
pyparsing==2.4.2          # via packaging
pytest-cov==2.7.1
pytest==5.1.0
pytz==2019.2              # via google-api-core
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
six==1.12.0               # via absl-py, astroid, google-api-core, google-auth, google-resumable-media, grpcio, h5py, keras-preprocessing, packaging, protobuf, tb-nightly, tensorflow
snowballstemmer==1.9.0    # via pydocstyle
tb-nightly==1.14.0a20190603  # via tensorflow
tensorflow==2.0.0-beta1
termcolor==1.1.0          # via tensorflow
tf-estimator-nightly==1.14.0.dev2019060501  # via tensorflow
toml==0.10.0              # via black
typed-ast==1.4.0          # via astroid, mypy
typing-extensions==3.7.4  # via mypy
urllib3==1.25.3           # via requests
wcwidth==0.1.7            # via pytest
werkzeug==0.15.5          # via flask, tb-nightly
wheel==0.33.6             # via tb-nightly, tensorflow
wrapt==1.11.2             # via astroid, tensorflow
zipp==0.5.2               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.1.0        # via google-api-core, markdown, protobuf, tb-nightly
